+++ pytest_common_subject/plugin.py
@@ -1,0 +1,5 @@
+def pytest_configure(config):
+    config.addinivalue_line(
+        'markers',
+        'precondition: mark a fixture to be evaluated before common_subject is invoked',
+    )